implement jwt auth for user

+++ migrations/2025-05-17-151639_create_users_table/up.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS users (
     email VARCHAR(255) NOT NULL UNIQUE,
     name VARCHAR(255) NOT NULL,
     password VARCHAR(255) NOT NULL,
-    balance_in_rs BIGINT NOT NULL DEFAULT 0,
+    balance_in_rs DOUBLE PRECISION NOT NULL DEFAULT 0.0,
     created_at TIMESTAMP NOT NULL DEFAULT now(),
     last_modified_at TIMESTAMP NOT NULL DEFAULT now()
 );
